vdb/requirements.txt
weaviate-client
pandas
python-dotenv
streamlit
